extract OOPanel
